refactor(sentry): enable performance tracing
