+++ .changeset/spicy-actors-roll.md
@@ -1,0 +1,6 @@
+---
+'@commercetools-backend/loggers': patch
+'@commercetools-frontend/sentry': patch
+---
+
+Enable performance tracing for Sentry
