+++ packages-backend/loggers/package.json
@@ -27,7 +27,7 @@
   "dependencies": {
     "@babel/runtime": "^7.19.0",
     "@babel/runtime-corejs3": "^7.19.0",
-    "@sentry/node": "7.16.0",
+    "@sentry/node": "7.19.0",
     "@types/triple-beam": "1.3.2",
     "express-winston": "4.2.0",
     "fast-safe-stringify": "2.1.1",
