+++ packages/sentry/package.json
@@ -20,15 +20,17 @@
   "dependencies": {
     "@babel/runtime": "^7.19.0",
     "@babel/runtime-corejs3": "^7.19.0",
+    "@commercetools-frontend/browser-history": "21.19.0",
     "@commercetools-frontend/constants": "21.19.0",
-    "@sentry/browser": "7.14.0",
+    "@sentry/react": "7.19.0",
+    "@sentry/tracing": "7.19.0",
     "@types/prop-types": "^15.7.5",
     "@types/react": "^17.0.49",
     "prop-types": "15.8.1"
   },
   "devDependencies": {
     "react": "17.0.2",
-    "sentry-testkit": "4.1.0",
+    "sentry-testkit": "5.0.3",
     "wait-for-expect": "3.0.2"
   },
   "peerDependencies": {
